Rename `requestAnimationTimeout` to avoid conflicts
